Moved RandomUI to random_sys_id_ui.py
